requirements.txt
asgiref==3.5.2
autopep8==1.6.0
boto3==1.24.19
botocore==1.27.19
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
django-cors-headers==3.12.0
django-environ==0.8.1
django-stubs==1.11.0
django-stubs-ext==0.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
jmespath==1.0.1
mypy==0.950
mypy-extensions==0.4.3
Pillow==9.1.1
psycopg2==2.9.3
pycodestyle==2.8.0
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
types-pytz==2021.3.8
types-PyYAML==6.0.7
typing_extensions==4.2.0
urllib3==1.26.9
